+++ changelog.md
@@ -291,6 +291,11 @@
 - 删除无用的钩子函数（已被RT-Thread钩子函数接管）
 - 修复`OSTimeDlyResume`函数中`p_tcb->TaskState`状态处理的问题
 - 将空闲任务纳入到`OS_CFG_STAT_TASK_EN`控制范围内
+- 删除`os_dbg.c`无用代码
+
+### 2020-10-06
+
+- 删除`os.h`无用代码
 
 
 
@@ -497,11 +502,15 @@
 
 对齐并整理注释
 
-删除无用的钩子函数（已被RT-Thread钩子函数接管）
+**[del]** 删除无用的钩子函数（已被RT-Thread钩子函数接管）
 
 **[bug fix]** 修复`OSTimeDlyResume`函数中`p_tcb->TaskState`状态处理的问题
 
 将空闲任务纳入到`OS_CFG_STAT_TASK_EN`控制范围内
+
+**[del]** 删除`os_dbg.c`无用代码
+
+**[del]** 删除`os.h`无用代码